apps/ai-service/pyproject.toml
[tool.poetry]
name = "ai-service"
version = "0.1.0"
description = "Operation Obsidian AI Service - Predictive Maintenance"
authors = ["Pocket Ops Team"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.24.0"
pandas = "^2.1.0"
scikit-learn = "^1.3.0"
psycopg2-binary = "^2.9.0"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
